@@ -4,19 +4,19 @@ version: 1.1.0+9
 homepage: https://github.com/PROGrand/flutter_qa
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: ">=2.18.1 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  rect_getter: ^0.1.0
+  rect_getter: ^1.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  test: '>=1.0.0 <2.0.0'
-  pedantic: ">=1.8.0+1 <2.0.0"
-  coverage: ">=0.13.0 <0.14.0"
+  test: ^1.22.0
+  pedantic: ^1.11.1
+  coverage: ^1.6.3
 
 flutter:
