@@ -18,7 +18,7 @@ List<String> destinations = [
 ///////////////////////////////////////////
 
 class MatchingPage extends StatefulWidget {
-  MatchingPage({Key key}) : super(key: key);
+  MatchingPage({Key? key}) : super(key: key);
 
   @override
   State<MatchingPage> createState() {
@@ -36,79 +36,89 @@ class MatchingPageState extends State<MatchingPage> {
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: <Widget>[
-        SizedBox(
-          height: 8,
-        ),
-        Text('Lorem ipsum dolor sit amet.',
-            style: Theme.of(context).textTheme.title),
-        MatchingWidget(
-          key: matchingKey,
-          builder: MatchingWidgetBuilder(
-              activeColor: Theme.of(context).primaryColor.withOpacity(0.25),
-              connectedColor: Theme.of(context).primaryColor,
-              onAddConnection: (s, d) {
-                connections[s] = d;
-              },
-              onRemoveConnection: (s) {
-                connections.remove(s);
-              },
-              onClearAll: () {
-                connections.clear();
-              },
-              sourcesCount: sources.length,
-              destinationsCount: destinations.length,
-              build: (BuildContext context, bool query, int index) =>
-                  (query ? _query(index) : _answer(index)),
-              connections: connections),
-        ),
-        RaisedButton(
-          child: Text('Clear'),
-          onPressed: () {
-            matchingKey.currentState.clear();
-          },
-        ),
-        SizedBox(
-          height: 16,
-        )
-      ],
+    return Material(
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: <Widget>[
+          SizedBox(
+            height: 8,
+          ),
+          Text('Lorem ipsum dolor sit amet.'),
+          MatchingWidget(
+            key: matchingKey,
+            builder: MatchingWidgetBuilder(
+                activeColor: Theme.of(context).primaryColor.withOpacity(0.25),
+                connectedColor: Theme.of(context).primaryColor,
+                onAddConnection: (s, d) {
+                  connections[s] = d;
+                },
+                onRemoveConnection: (s) {
+                  connections.remove(s);
+                },
+                onClearAll: () {
+                  connections.clear();
+                },
+                sourcesCount: sources.length,
+                destinationsCount: destinations.length,
+                build: (BuildContext context, bool query, int index) =>
+                    (query ? _query(index) : _answer(index)),
+                connections: connections),
+          ),
+          ElevatedButton(
+            child: Text('Clear'),
+            onPressed: () {
+              matchingKey.currentState?.clear();
+            },
+          ),
+          SizedBox(
+            height: 16,
+          )
+        ],
+      ),
     );
   }
 
   Widget _query(int index) {
     return Padding(
-        padding: EdgeInsets.only(top: 8, left: 16, bottom: 8),
-        child: Container(
-            padding: EdgeInsets.all(16),
-            color: const Color(0xffe4f2fd),
-            foregroundDecoration: BoxDecoration(
-                border: Border.all(
-              color: const Color(0xffc2d2e1),
-              width: 2,
-            )),
-            child: Center(
-              child: Text('${sources[index]}',
-                  style: TextStyle(color: Colors.black, fontSize: 12.0)),
-            )));
+      padding: EdgeInsets.only(top: 8, left: 16, bottom: 8),
+      child: Container(
+        padding: EdgeInsets.all(16),
+        color: const Color(0xffe4f2fd),
+        foregroundDecoration: BoxDecoration(
+          border: Border.all(
+            color: const Color(0xffc2d2e1),
+            width: 2,
+          ),
+        ),
+        child: Center(
+          child: Text(
+            '${sources[index]}',
+            style: TextStyle(color: Colors.black, fontSize: 12.0),
+          ),
+        ),
+      ),
+    );
   }
 
   Widget _answer(int index) {
     return Padding(
-        padding: EdgeInsets.only(top: 8, right: 16, bottom: 8),
-        child: Container(
-          padding: EdgeInsets.all(16),
-          color: const Color(0xffe4f2fd),
-          foregroundDecoration: BoxDecoration(
-              border: Border.all(
+      padding: EdgeInsets.only(top: 8, right: 16, bottom: 8),
+      child: Container(
+        padding: EdgeInsets.all(16),
+        color: const Color(0xffe4f2fd),
+        foregroundDecoration: BoxDecoration(
+          border: Border.all(
             color: const Color(0xffc2d2e1),
             width: 2,
-          )),
-          child: Center(
-            child: Text('${destinations[index]}',
-                style: TextStyle(color: Colors.black, fontSize: 12.0)),
           ),
-        ));
+        ),
+        child: Center(
+          child: Text(
+            '${destinations[index]}',
+            style: TextStyle(color: Colors.black, fontSize: 12.0),
+          ),
+        ),
+      ),
+    );
   }
 }
